Resize post instantly iff clicking resize handle.
Max width and height now removed instantly only when *clicking*
the resize-e handle. In all other cases, size constraints are removed
on jQuery's resize-start event. So, dragging handle works as usual.
But if you click it, the post i instantly resized.

Click the resize-e(ast) handle, and the post expands eastwards.
Click the south/southeast handles, and it expands downwards.