Outline new replies.
Set a cookie and/or a javascript variable to the last timestamp
of the discussion. (In LayoutManager). Outline threads created
more recently than that timestamp. (In debiki.js).

Send the current page timestamp to the server (via a cookie or
a hidden parameter -- both, currently, weird), so the server
can send it back, as the above-mentioned last-page-timestamp.

Also:
 - Hide LayoutManager.menus
 - Add config class LayoutVariables
 - Add Debate.lastChangeDate
 - Add css class 'dw-new'.