Load Edit:s from yaml file
Add Edit class, and Debate.edits. Use a new concise DebikiMapConstr2
construction to load Edit instances from a yaml file.

Storing the edited text, or the edit steps? --- Right now, the whole
edited text is stored in the yaml file.  Another approach is to store
only the edit steps that transform the original text to the edited text.
I.e. some kind of diff. e.g. change-that-line, delete-that-paragraph.
This could be represented (in Scala) like so:

   Edit.editSteps: List[EditSteps] // instead of Edit.text
   sealed class EditStep
   object EditStep {
     case class ReplaceAll(newText: String) extends EditStep
     case class ReplaceWord(...) extends EditStep
     case class MoveParagraph(...) extends EditStep
     // ...
   }

However, I think it's better to store the edited text only, but no
transformation steps. If only the end result is remembered, then, in the
future, I can use any good diff algorithm to calculate the
transformation steps, and understand what changes were intended, and
apply these changes to the post -- also if it's been edited by other
persons inbetween. (This would be hard to do, if I settled for a
edit-step-representation today -- it would be no good representation.)