Move css and js files to toserve/ subdir
Debiki-core is a JAR dependency of Lift-Web app debiki.com, and Lift-Web
requires that resource files located inside JAR archives be placed in a
dedicated directory, so resource names can't be spoofed. So move the
resources to a "toserve/" sub directory --- this is the resource
directory Lift-Web uses, by default.

Update App.scala to take into account the new resource path.