Keep post text selection until inline menu click
Currently, when you select text in a post, the menu item pops up, and
directly finds the text at the start/end of the selection, but this
destroys the selection!  So calculate the start/end lazily — on menu
item click —, and splitBodyPlaceInlines lazily too.  This avoids
destroying the selection before menu item click.  Also make a copy of
the selection, because the real selection is modifieid/destroyed on a
menu item click.