Tweak root post edit form CSS
Don't restrict root post width, and make the root post preview as wide
as the root post.  But now it becomes too wide! when Twitter's Bootstrap
isn't in use (I'm testing it right now, building a home page). Fix in
the next commit?

Wrap the root post preview in a .dw-p-bdy-blk, like in a real post, and:
- Remove some no longer needed .dw-ed-tab-preview styles
- Remove pointless `div.' in `div.dw-p-bdy-blk'
- Add some missing edit <form> styles matching .dw-p-bdy