Delete old unused class EditVote
Instead of using a separate EditVote class, in the future, if needed,
allow people to rate all *actions*, not only Posts.  Then, an Edit
suggestion that have been rated so that it has a high fitness for
Interesting/Useful/+1 tags, could be auto-applied, or the author (which
can apply edits) could be notified.