Stop using SVG Web
Loading svg.js takes long, and in $createSvgRoot, appendChild(new-
svg-elem) takes rather long (according to an old comment of mine)
because of  _processSVGScript, in svg.js.

In Chrome, this speeds up $createSvgRoot with a factor 10?
From these run times: 57, 42, 37, 14, 52, 18, 25, 42, 41
To these: 7, 1, 6, 1, 3, 15, 6, 1, 3, 1, 2
(Google Chrome 17, a page with 43 replies.)

In FF, did this speed up $createSvgRoot 4 times (from 82ms down to
22ms). Or did it make $createSvgRoot relatively slower? I guess `Time'
but not `%' or `Own-Time' is interesting, because *with* SVG Web much
time is spend in function in svg.js, and I'd guess `%' does not include
that time.  `Time' does include it though.

                  Calls %       Own-Time Time      Avg      Min      Max
- With SVG Web, 5 reloads:
  $createSvgRoot  88    0.38%   2.352ms  82.543ms  0.938ms  0.796ms  4.596ms
- Without SVG Web, 5 reloads:
  $createSvgRoot  88    1.48%   7.579ms  22.21ms   0.252ms  0.201ms  1.439ms

- arrowFromThreadToReply also faster? With SVG Web:
  arrowFromThreadToReply  45  0.88%  5.457ms  82.805ms  1.84ms  0.386ms  4.263ms
- Without: (avg. 1.84 --> 1.098)
  arrowFromThreadToReply  45  0.93%  4.775ms  49.4ms    1.098ms 0.38ms   2.977ms