Scroll modal dialogs into view, nice on mobiles. Crazy commit!
This crazy commit includes lots of debug code and an insane workaround
for an Android bug, that makes certain Android versions report the wrong
screen.width, window.innerWidth, document.documentElement.clientWidth
and heights, after Android has auto-zoomed into a text input.  (When you
focus a text input, Android auto-zooms that input (and shows the
keyboard).)

The workaround does not always work though, because it's not possible to
detect all cases when a text input is focused (and Androdi auto zooms
and triggers the bug). So I'll remove it in the next commit?