Load debiki.js files synchronously, or bugs happen randomly
Unfortunately: Modernizr has to be loaded and evaluated before the sync
loading of debiki.js files starts. I've checked Chrome's dev tools
timeline, and on my computer the loading of combined-debiki-*.min.js is
delayed with 70ms to 500ms depending on randomness and whether you do a
refresh or a complete page reload. (Localhost server.)