Fix CachingTenantDao bug: Don't forget people who saved page
Don't override createPage() to reload and cache the saved page, because the
reloaded page contains no info on the one that created it (unless some actions
were saved as well), and then there's a None.get error later (in AppEdit).

Also let savePageActions() remove the page from the cache, instead of
potentially reloading it twice.