Add /favicon.ico route and file
At least hereafter no more browsers will ask for /favicon.ico again and
again and again.

About favicon-test-2.ico: (currently in use)
The blue background is the sky; the white stuff is a white rose. This is
in memory of The White Rose, http://en.wikipedia.org/wiki/White_Rose,
which was a non-violent resistance group operating in Nazi Germany.
They became known for an anonymous leaflet and graffiti campaign that
called for active opposition to dictator Adolf Hitler's regime.
The "D" is the first letter in Debiki.

About favicon-test-1.ico:
This is supposed to be a white leaf of a white rose, however the left
part of the leaf has been cut off. So only the right part remains. And
the right part resembles a "D", actually.