Assume Vagrant script is run from parent directory
That parent directory is supposedly a site specific Git repo, namely one
derived from debiki-site-seed. Vagrant is to be started from that
directory, so that a certain themes/ folder is shared with the Vagrant
VM.

In the parent dir, the debiki-server subdir should be named "server".
So prefix some paths with "server/".

Also have the bootstrap script CD to /vagrant/server/, so the user
doesn't need to do this.