Re-add modules/debiki-dao-rdb
The problem seems to have been that I 'cloned' debiki-dao-pgsql by copying it
(`cp -a ...`) but then the '.git' file was copied, of course, and it's contents
was:
  ../../../.git/modules/server/modules/modules/debiki-dao-rpgsql
rather than:
  ../../../.git/modules/server/modules/modules/debiki-dao-rdb
so when I deleted 'debiki-dao-pgsql', the repo become corrupted (parts of it
suddenly missing).

(Perhaps I need not have added all these commits to fix that problem.)

Also fix path to modules/debiki-dao-rdb in Build.scala.