Fix parse-empty-JSON bug
If loading page when not logged in, no user-page-data will be included
in .dw-user-page-data. But if then going to another page, logging in,
and using the browser's back button to navigate back, again no user-page-
data is present, but this time the user is logged in and this results
in an attempt to parse the empty string. So don't check for the
presence of the element only, also check the text length.