Add broken unfinished scrypt server relief stuff
Client side scrypt-hashing is way too slow on my mobile phone (less than
1 year old), so let's do stuff server side instead.

However, in the future, can have the client do the computations, if the
server is under heavy load. (And store the sha-256 of the scrypt derived
key server side. And usually have the server compute the derived key and
sha-256-hash it.)