Work around Chrome 63 same-origin bug
Chrome fails with a cross origin error, although is same origin,
if, from an iframe, one opens a popup, in the popup goes to Google's
login, and then in the popup back to the iframe's origin and attempt
to access window.opener.debiki from the popup. The popup is now back
at the same origin, still Chrome 63 says cross-origin-error.