Update license and copyright, mention BUSL
And rename AGPL file so GitHub finds it.

Don't mention Discourse — apparently mentioning Discourse made some
people believe that Talkyard was a fork of Discourse, but it very much
is not: I wrote the first lines in what eventually became Talkyard, in
early 2010, but Discourse wasn't started until 2012 and was announced
publicly first in 2013.