Persona mode, better anonymous comments
Make it possible to enter Anonymous mode or Self mode, to always be
anonymous if allowed, or always oneself, if allowed.  If unclear which
persona to use (e.g. if to reply as oneself, or anonymously) open a
dialog and ask the user what they want.

Choose alias when editing titles too, not just OP & comments.

Show any selected or default persona (e.g. anon, in anon by default
cats), next to one's username. Send to server in X-Ty-Persona header.
Make the server reject requests if any alias (i.e. a persona other than
oneself) selected (e.g. an anonym), but the endpoint doesn't support
aliases.

Show a "not anonymous" tips if switching from an anon to a not-anon
category when composing a new topic.

And:
- Cache discussion properties, in the browser, when loading a new page.
  So won't need to recompute at different locations (in the code) and
  won't accidentally do it differently.
- In Authz, check if anons allowed: Category settings, and anonOnPageId.
- Fix: Add lazy-created anon voter to knownAnons.
- Remove `[get_anon]` dupl code.