Optimize: Don't emit change event
Don't emit, if notification fns didn't do anything: When bookmark-link
navigating to a page that's already in the cache, then, after
markAnyNotificationAsSeen() had run, ReactStore.emitChange() gets
called, and all change event handlers take about 100ms to complete (in
total, together, in a VM in my slooow Core i5 laptop). That's a lot!
Pretty important to avoid emitting events when not needed.

And don't emit event, if just scrolling to & highlighting a comment.
Currently, scrollAndShowPost() emits a store event that takes about 100
ms to handle in Chrome in a VM on my slooow laptop. But when not
emitting the event, takes less than 0.1 ms. — Emitting an event, is only
needed if the comment gets lazy-loaded or other comments (ancestors)
need to be uncollapsed for it to be shown.