Remove old forms, and last traces of LiftWeb
This old not-in-use maybe-good-to-have (I though) code causes compilation
errors now with Scala 2.13 and Play 2.9. Something related to
immutable/collection.Seq but not so easy to fix. Remove it.

Now old LiftWeb's Html5Writer is not longer needed, remove too.

Remove nu.validator.htmlparser too, was needed by LiftWeb only.