Send a more detailed patch to post author
Generate two React store patches: One for the post author / voter, and
another to send over WebSocket to others who can see the page.

Then, with info about which anonyms are the author's own anonyms,  it'll
be possible to update the UI directly so the author sees the text
"(you)" next to hans anon posts — otherwise needs to reload the page.

Related:
  a4de45fd7937  Fix anonymity bug: Don't incl anonForId in WebSocket msgs.